+++ integration/setup/pfs-entrypoint.sh
@@ -9,6 +9,7 @@ export ETH_RPC="http://localhost:8545"
 
 /opt/services/venv/bin/python -m pathfinding_service.cli \
     --accept-disclaimer \
+    --host 0.0.0.0 \
     --keystore-file /opt/services/keystore/UTC--2020-03-11T15-39-16.935381228Z--2b5e1928c25c5a326dbb61fc9713876dd2904e34 \
     --token-network-registry-contract-address "${TOKEN_NETWORK_REGISTRY_ADDRESS}" \
     --user-deposit-contract-address "${USER_DEPOSIT_ADDRESS}" \