+++ raiden-dapp/src/views/BackupState.vue
@@ -20,7 +20,10 @@
           </div>
         </v-list-item-content>
       </v-list-item>
-      <!-- <v-list-item class="backup-state__buttons__upload-state">
+      <v-list-item
+        class="backup-state__buttons__upload-state"
+        @click="uploadState = true"
+      >
         <div class="backup-state__buttons__upload-state__icon">
           <v-img :src="require('../assets/state_upload.png')"></v-img>
         </div>
@@ -29,13 +32,15 @@
             {{ $t('backup-state.upload') }}
           </div>
         </v-list-item-content>
-      </v-list-item> -->
+      </v-list-item>
     </v-list>
     <download-state-dialog
       :visible="downloadState"
       @cancel="downloadState = false"
     >
     </download-state-dialog>
+    <upload-state-dialog :visible="uploadState" @cancel="uploadState = false">
+    </upload-state-dialog>
   </div>
 </template>
 
@@ -43,16 +48,19 @@
 import { Component, Vue } from 'vue-property-decorator';
 import { mapGetters } from 'vuex';
 import DownloadStateDialog from '@/components/DownloadStateDialog.vue';
+import UploadStateDialog from '@/components/UploadStateDialog.vue';
 
 @Component({
   components: {
-    DownloadStateDialog
+    DownloadStateDialog,
+    UploadStateDialog
   },
   computed: { ...mapGetters(['isConnected']) }
 })
 export default class BackupState extends Vue {
   isConnected!: boolean;
   downloadState: boolean = false;
+  uploadState: boolean = false;
 }
 </script>