+++ raiden-dapp/CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
 ## Unreleased
+### Added
+- [#1222] Enables state upload
 
 ### Fixed
 - [#1310] Disable state download if not connected to web3 provider.
 
+[#1222] https://github.com/raiden-network/light-client/issues/1222
 [#1310]: https://github.com/raiden-network/light-client/issues/1310
 
 ## [0.5.2] - 2020-04-07
@@ -19,6 +22,7 @@
 - [#1243] Fix width of transfer input on mobile screens.
 - [#1242] Fix transfer modal not closing instantly.
 
+[#694] https://github.com/raiden-network/light-client/issues/694
 [#694]: https://github.com/raiden-network/light-client/issues/694
 [#1243]: https://github.com/raiden-network/light-client/issues/1243
 [#1242]: https://github.com/raiden-network/light-client/issues/1242